Fixed the strange aborting issue!
Just had to re-work how ctypes was passing data around. (After lots of banging head against the desk action). Was able to receive correct location data from each beacon and verify it against the dashboard (data is sent little endian!). Next step is to clean up the locaiton processing, get it into a loop, and present the data to the rest of the system in a useful API.